% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{align_sets}
\alias{align_sets}
\title{Align set of query epitope sequences against set of reference epitope sequences}
\usage{
align_sets(query.set, ref.set, env = NULL, num.cores = 2,
  aln_matrix = "BLOSUM62")
}
\arguments{
\item{query.set}{A matrix of Amino Acid epitope sequences to align against the reference epitope sequences.}

\item{ref.set}{A matrix of Amino Acid epitope sequences, epitopes with names in rows and bases by position in columns.}
}
\value{
A matrix of alignment scores for query epitopes against reference epitopes
}
\description{
Aligning set of query epitopes by multi-threaded parallel execution of align_to_red_epitopes()
function. The alignment score between all pairs of query and ref epitopes is reported as a matrix.
}
\examples{
\dontrun{
align_sets(query.set, ref.set, env=NULL, num.cores=2)
}
}
\keyword{internal}
