% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{align_to_ref_epitopes}
\alias{align_to_ref_epitopes}
\title{Alignment of singular epitope vector against a reference matrix of epitope, rows as epitopes with names and bases in 
columns by position.}
\usage{
align_to_ref_epitopes(x, ref, mag = 4, aln_matrix = "BLOSUM62")
}
\arguments{
\item{x}{A character vector of Amino Acid bases.}

\item{ref}{A matrix of Amino Acid epitope sequences, epitopes with names in rows and bases by position in columns}

\item{mag}{Numeric value to use for seq() function to create distant sequence of weights}
}
\value{
A numeric vecdtor of scores for x alignment to the each epitope in ref
}
\description{
Align the epitope sequence to the reference set of epitope sequences. Sequences are alignment globally
from end to end with matches assigned a postitive score as the position specific weight and mis-matches
are assigned a negative score as the negative of the position specific weight. Final alignment score is
the sum of all the positive and neagative scores from each base. This ensures that the alignment is matches
in the middle are scored higher than the alignments with mis-matches in the middle.
}
\examples{
\dontrun{
align_to_ref_epitopes(x, ref, mag=4)
}
}
\keyword{internal}
