% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{epitopeDF_to_AAmat}
\alias{epitopeDF_to_AAmat}
\title{Convert the epitopes data frame to amino acid matrix with bases separated into columns}
\usage{
epitopeDF_to_AAmat(sequence.name, epitope.DF)
}
\arguments{
\item{sequence.name}{Name of the input protein sequence for predicting epitopes.}

\item{epitope.DF}{Data frame of the epitopes predicted by the IEDB server.}
}
\value{
A matrix of Amino Acid epitope sequences, epitopes with names in rows and bases by position in columns
}
\description{
Data frame of predict epitopes is converted to a matrix of epitope sequences by rows and their
bases in columns.
}
\examples{
\dontrun{
epitopeDF_to_AAmat(sequence.name, epitope.DF)
}
}
\keyword{internal}
