% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{fetch_alleles}
\alias{fetch_alleles}
\title{Fetch the all MHC alleles for a specified species from the IEDB server API}
\usage{
fetch_alleles(species = "human",
  api.url = "http://tools-api.iedb.org/tools_api/mhci/")
}
\arguments{
\item{species}{Common name of the species to query the IEDB server.}

\item{api.url}{URL of the IEDB server API, default api is for MHC-1 prediction.}
}
\value{
A vector of the MHC allele associated with the specified species
}
\description{
Pass the species information to IEDB server for fetching all the associated MHC alleles.
}
\examples{
\dontrun{
fetch_alleles(species="human", api.url="http://tools-api.iedb.org/tools_api/mhci/")
}
}
\keyword{internal}
