% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{fetch_epitopes}
\alias{fetch_epitopes}
\title{Fetch the MHC-1 eptiopes prediction result from the IEDB server API}
\usage{
fetch_epitopes(sequence, method = "recommended", allele = "H-2-Kb",
  length = 9, api.url = "http://tools-api.iedb.org/tools_api/mhci/")
}
\arguments{
\item{sequence}{Protein sequence from which the epitopes are predicted.}

\item{method}{MHC class 1 binding prediction methods.}

\item{allele}{MHC allele.}

\item{length}{Length of the peptide to be predicted.}

\item{api.url}{URL of the IEDB server API, default api is for MHC-1 prediction.}
}
\value{
A data.frame of the IEDB MHC-1 prediction result on the input sequence
}
\description{
Pass the protein sequence to IEDB server for predicting MHC-1 binding epitopes.
The predicted epitope sequences are pasrsed and formatted as a matrix with 
sequence in rows and each base of the sequence as a column.
}
\examples{
\dontrun{
fetch_epitopes(sequence, 
   method="recommended", 
   allele="H-2-Kb", 
   length=9, 
   api.url="http://tools-api.iedb.org/tools_api/mhci/")
}
}
\keyword{internal}
