% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{filter_epitopes}
\alias{filter_epitopes}
\title{Filter the epitopes by IC50 score}
\usage{
filter_epitopes(epitope.DF, ic50.threshold)
}
\arguments{
\item{epitope.DF}{Data frame of the epitopes predicted by the IEDB server.}

\item{ic50.threshold}{A numerical threshold value for the IC50 score of epitope.}
}
\value{
A data.frame of the predicted epitopes filtered by the IC50 threshold score
}
\description{
Filter the predicted epitope sequence data frame by the user provided IC50 score threshold.
The epitopes with IC50 score <= IC50 threshold are selected and returned.
}
\examples{
\dontrun{
filter_epitopes(epitope.DF, ic50.threshold)
}
}
\keyword{internal}
