% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{get_best_alignment}
\alias{get_best_alignment}
\title{Get the pairs of query and reference epitopes with the best alignment score or above the 
specified alignment score threshold}
\usage{
get_best_alignment(alnRes, query.set, ref.set, best = TRUE, score = NULL)
}
\arguments{
\item{alnRes}{A matrix of alignment scores for query epitopes against reference epitopes.}

\item{query.set}{A matrix of Amino Acid epitope sequences to align against the reference epitope sequences.}

\item{ref.set}{A matrix of Amino Acid epitope sequences, epitopes with names in rows and bases by position in columns.}

\item{best}{A boolean flag to switch between selecting the pairs of epitopes with best alignment score or use the 
input score value as a threshold.}

\item{score}{A numerical value to use as alignment score threshold.}
}
\value{
A data frame of alignment result for the selected pairs of aligned epitopes
}
\description{
Aligning set of query epitopes by multi-threaded parallel execution of align_to_red_epitopes()
function. The alignment score between all pairs of query and ref epitopes is reported as a matrix.
}
\examples{
\dontrun{
get_best_alignment(alnRes, query.set, ref.set, best=TRUE, score=NULL)
}
}
\keyword{internal}
