% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{get_pos_weights}
\alias{get_pos_weights}
\title{Get position specific weights, higher scores for the mid bases compared to tails.}
\usage{
get_pos_weights(x, mag = 4)
}
\arguments{
\item{x}{A character vector of Amino Acid bases.}

\item{mag}{Numeric value to use for seq() function to create distant sequence of weights}
}
\value{
A numeric vector of scores equal to the length of the input vector of Amino Acids
}
\description{
Compute the position specific scores for the amino acid bases with heavier weight for the middle bases 
and lower score for bases towards the edges. Scores are a function of amino acid bases vector length
and the magnitude of score elevation by base distance from edges, bases on the third quartile of edges
are simplified to magnitude of 1 beginning from 1.
}
\examples{
\dontrun{
get_pos_weights(x=AA.vec, mag=4)
}
}
\keyword{internal}
