% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{predict_epitopes}
\alias{predict_epitopes}
\title{Predict the MHC-1 eptiopes from the protein sequence by using the IEDB server API}
\usage{
predict_epitopes(sequence.name, sequence, method = "recommended",
  allele = "H-2-Kb", length = 9,
  api.url = "http://tools-api.iedb.org/tools_api/mhci/")
}
\arguments{
\item{sequence.name}{Name of the input protein sequence for predicting epitopes.}

\item{sequence}{Protein sequence from which the epitopes are predicted.}

\item{method}{MHC class 1 binding prediction methods.}

\item{allele}{MHC allele.}

\item{length}{Length of the peptide to be predicted.}

\item{api.url}{URL of the IEDB server API, default api is for MHC-1 prediction.}
}
\value{
A matrix of Amino Acid epitope sequences, epitopes with names in rows and bases by position in columns
}
\description{
Pass the protein sequence to IEDB server for predicting MHC-1 binding epitopes.
The predicted epitope sequences are parsed and formatted as a matrix with 
sequence in rows and each base of the sequence as a column.
}
\examples{
\dontrun{
predict_epitopes(sequence.name="Seq1", 
     sequence, method="recommended", 
     allele="H-2-Kb", 
     length=9, 
     api.url="http://tools-api.iedb.org/tools_api/mhci/"
)
}
}
\keyword{internal}
