% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPItOMe.R
\name{read_epitope_seq}
\alias{read_epitope_seq}
\title{Read the epitope amino acid sequences from the fasta file}
\usage{
read_epitope_seq(iFile)
}
\arguments{
\item{iFile}{Path to a file containing epitope sequences in fasta format.}
}
\value{
A matrix of Amino Acid epitope sequences, epitopes with names in rows and bases by position in columns
}
\description{
Read epitope sequences from a fasta file, provided with the correct path to the file.
The read sequences are provided as a matrix with sequence in rows and each base of the
sequence as a column.
}
\examples{
\dontrun{
read_epitope_seq(iFile=input.file)
}
}
\keyword{internal}
